# Simplify FASTQWriter