This simplifies the write method. This is slightly *slower*

Before: 
```
Benchmark #1: python dnaio_read_and_write.py ~/test/big2.fastq /dev/null
  Time (mean ± σ):      1.957 s ±  0.025 s    [User: 1.849 s, System: 0.108 s]
  Range (min … max):    1.916 s …  1.986 s    10 runs
```

After:
```
Benchmark #1: python dnaio_read_and_write.py ~/test/big2.fastq /dev/null
  Time (mean ± σ):      2.056 s ±  0.057 s    [User: 1.939 s, System: 0.117 s]
  Range (min … max):    2.010 s …  2.174 s    10 runs
```

However in terms of throughput it is not that big of a difference, and there is always the option of using xopen, and using the resulting file and `fastq_bytes` directly:

```
Benchmark #1: python dnaio_read_and_write.py ~/test/big2.fastq /dev/null
  Time (mean ± σ):      1.566 s ±  0.026 s    [User: 1.455 s, System: 0.110 s]
  Range (min … max):    1.521 s …  1.609 s    10 runs
```

This is much faster than FastqWriter, and it is simply impossible for FastqWriter to reach those speeds anyway. Simplification increases its overhead from 400ms to 500ms, but makes the code much easier to read.