@marcelm 2022-02-16
Simplification at the cost of speed hasn’t really been the aim of your previous PRs, why now? Cutadapt supports writing FASTA and FASTQ files, which is why it doesn’t use `fastq_bytes` directly and instead goes through the `FastqWriter` (or a `FastaWriter`).