@rhpvorderman 2022-02-16
> Simplification at the cost of speed hasn’t really been the aim of your previous PRs, why now? 

I figured that since FastqWriter is slower than using xopen directly anyway, we might just as well make it simple. But it is only 20 lines of code, so I guess this PR is redundant. I will close this and I apologize for bringing this up.