@marcelm 2022-02-16
No worries. I did not have it on the radar that FastqWriter may be slow, so that is good to know.

It’s a bit surprising to me that the overhead is so relatively large since it’s "just" an extra function call and two attribute lookups (-ish). I tried just now to move the `writers.py` into a `writers.pyx` module and that didn’t help too much, so I’ll leave that be for the moment.