@rhpvorderman 2022-02-16
> It’s a bit surprising to me that the overhead is so relatively large since it’s "just" an extra function call and two attribute lookups (-ish). I tried just now to move the writers.py into a writers.pyx module and that didn’t help too much, so I’ll leave that be for the moment.

Well, this is tested on exactly 5 million recrods, so it is only 80 ns per record. An attribute lookup is roughly 20 ns and a function call is roughly 40 ns, so two attribute lookups and the extra function indeed add up to that amount.

So that's why I think that FastqWriter cannot be improved much. Sure you can eliminate one of the attribute lookups maybe by aliasing `self._file.write` to `self._write`, but that is about it. The wrapping cost is already very low at 80 ns. The problem is that FASTQ Illumina records are extremely small and therefore very numerous.