@marcelm 2022-02-16
> Sure you can eliminate one of the attribute lookups maybe by aliasing self._file.write to self._write, but that is about it.

Yepp, that was the first thing I tried an it indeed didn’t help much.