@rhpvorderman 2022-02-16
No worries, pysam can't write FASTQ anyway (I don't see it in the docs) and dnaio is faster than anything else in the python world. 
I benchmarked pysam reading too and ... well it is offtopic, but dnaio is at least 4 times faster. Also gzipped reading is about 3 times faster. So we really should write a paper.